Stub ECS ListTasks
Summary:
https://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ListTasks.html

The actual endpoint has a lot of different ways to filter the list
query. Our stub only filters by `cluster` and `family` because that's
all I anticipate initially needing to filter on in the EcsRunLauncher.

This will need to be extended to support paginated results.

Test Plan:
alangenfeld, dgibson, johann

Reviewers:

Subscribers: