[dagit] Switch to a new icon for jobs, limit overuse of pipeline icon
Summary:
Screenshots of changes below:

Only the change in the run list and the definition sidebar are not feature flagged - I think these icons should be changed for existing users too since the old iconography was a bit confusing. For the run table, the icon was the only thing clickable to reach the pipeline, so it makes sense to give it an action icon rather than the pipeline icon. For the sidebar, you're typically viewing the selected solid so the pipeline icon wasn't a great choice.

Test Plan: Review screenshots

Reviewers: dish

Differential Revision: https://dagster.phacility.com/D8394