Tag runs with ECS info
Summary:
By storing the ECS run id in run storage, we no longer need to list all
tasks to find one with a matching tag. We're still tagging the ECS task
with the dagster run id so it's clear why that task was created when
looking in AWS - eventually we might choose to add additional dagster
metadata as well.

Test Plan:
unit

Reviewers:

Subscribers: