Cross-reference ECS reference deployment
Summary:
This links back out to GitHub to show the full ECS reference deployment.
Alternatively, we could create a new deployment guide in our docs site
with the contents of the README.

I think the rest of the AWS content is otherwise worth keeping (the RDS
and S3 stuff) so I didn't want to replace it wholesale.

Test Plan:
none

Reviewers:
catherinewu, dgibson, johann

Subscribers: