[dagit] Fix fragment ID for runs
Summary:

Fix an issue where the Gantt chart doesn't update when a run goes from "queued" to "started". The issue here is that we're not properly referencing the PipelineRun object ID in the Apollo cache, so it's never updated when the status changes.

Test Plan:

Run the queued run coordinator daemon locally, start a run. Verify that the Gantt chart switches from the "Queued" state to a running state, without any weird delays.

Reviewers:

bengotow, dgibson

Subscribers: