[dagit] Don't show "Refresh config" if no partition selected
Summary:

Resolves #4296

If a partition name hasn't been selected for a config session, don't show the "Refresh config" button when reloading the workspace.

I think it doesn't make sense to allow regenerating the config via this button if no partition name is selected, because there's nothing to regenerate from -- the user would have to select a partition name anyway, which would perform the config generation.

Test Plan:

View playground config for a pipeline with a partition set. Create a new session tab and select a partition set, but do not choose a partition name. Reload the workspace, verify that other session tabs show the "Refresh config" button, but that the one without a selected partition name does not.

Reviewers:

bengotow, catherinewu, prha

Subscribers: