Merge Workspace into WorkspaceProcessContext, Make WorkspaceRequestContext an IWorkspace
Summary: Right now there is Workspace, WorkspaceProcessContext, and WorkspaceRequestContext. Would like to move us to a world where the thing that we currently call WorkspaceRequestContext is the 'workspace' (since we might start passing it into the run launcher, for example), and the thing that we call WorkspaceProcessContext is more like a factory for creating workspaces. In service of that goal, merge the first two into WorkspaceProcessContext

Test Plan: BK

Differential Revision: https://dagster.phacility.com/D8556