Merge explorer into the Overview tab, remove Definition tab for all users
Summary:
This diff consolidates the "Definition" and "Overview" tabs for all users. The pipeline's DAG is shown on the left, and the right sidebar shows the contents of the old overview tab. It also cleans up the pipeline explorer so it doesn't internally modify the explorerPath and can be mounted more flexibly.

I used a context to inject the additional tab into the sidebar when the pipeline explorer is used for the Overview tab. I set up a context rather than using props.children or passing the data down into the tree because I think it'd be nice to sprinkle more job- or run- related stuff into the DAG as well, but it's a bit awkward right now.

This diff also fixes the scrolling issue caused by styles on the graph pages.

Todo:

- The new tab is titled "Job" even for non feature-flagged users (waiting to see if this new third tab is here to stay.)

Test Plan: Give it a spin, run snapshots

Reviewers: dish

Differential Revision: https://dagster.phacility.com/D8614