style: adopt pre-commit for Python black formatting
Summary:
A contributor recommended that we adopt (pre-commit) for the formatting
steps that we usually test for in buildkite. We'll be migrating to Github
soon anyways, so adopting this toolchain will be great to replace what
arclint usually did for us.

Overall, pretty nice experience. `pip install pre-commit` and you're off to the
races.

I decided to format all python files rather than whatever we configured in the `Makefile`
before. Docs are formatted with line-length=78 for ease of reading.

Since this tool installs the hooks on its own, we remove dependencies of `black` from setup.

We'll use this tool to adopt the other linters we have as well.

Test Plan: buildkite, `pre-commit run black --all-files`

Reviewers: jordansanders, alangenfeld, max

Differential Revision: https://dagster.phacility.com/D8684