[dagit] Upgrade chart.js
Summary:

Upgrade chart.js to v3, which allows us to get past an existing security issue with v2 that makes it much harder to implement a CSP in internal, namely that v2 would require us to enable `unsafe-inline`.

This required changing some `options` objects for v3, but I think I did it correctly. It would be good to bang away at this in dev to make sure that the charts are all behaving as expected.

Test Plan:

View charts in Dagit, interact with them. Verify that things seem to be working normally, including:

- Zoom/pan behavior in tick history
- Hover/tooltip behavior

Reviewers:

bengotow, prha

Subscribers: