[dagit] Fix: Choose partition set for session tab
Summary:

Fix the case where there is no preset but one partition set in the Playground. Currently we are showing neither the partition set picker nor the partition picker, because the default session object has no partition set chosen.

To fix this, when there is only one partition set that matches the mode, initialize the session with that partition set name.

Test Plan:

Load playground for a job that has a partition set but no preset. Verify that the partition picker renders, with the red outline, and that config is populated correctly when choosing a partition.

Reviewers:

bengotow, sandyryza

Subscribers: