[dagit] Crag: Add "job" to Runs filter
Summary:

For crag-flagged users, allow filtering by "job" in Runs views (Runs page, Runs tab on Job page). The displayed jobs in the filter input are pipeline:mode tuples, or just the pipeline name when the mode is `default`.

This makes use of https://dagster.phacility.com/D8629, which added the `mode` column and GraphQL filter parameter. When a job is chosen in the filter input, the job name is turned into the appropriate GraphQL parameters (`pipelineName`, `mode`) for the request.

Test Plan:

With flag enabled:

- View a job, then its Runs tab. Verify that the permanent token is correct, and that runs are requested correctly.
- View the Runs page, filter by `job`. Verify that pipelines with only "default" mode display and query correctly, and that pipelines with modes do as well.

With flag disabled, verify that the behavior is unchanged from before: I can query by pipeline, and it works.

Reviewers:

bengotow, alangenfeld, cdecarolis

Subscribers: