[dagit] Crag: Don't show Partitions tab if none for Job
Summary:
For Jobs that do not have partitions, don't show the Partitions tab.

When determining whether to render the tab, find partition sets that share the current pipeline name (or job name) and, if flagged, also check the mode name in the URL. If there are any, show the tab.

Test Plan:
In flag, view a job in Dagit. Verify that the Partitions tab appears when there is a pipeline name/mode match. Verify that it does not appear otherwise.

Outside flag, repeat, verify that the tab appears when there is a pipeline name match.

Reviewers: bengotow, sandyryza

Differential Revision: https://dagster.phacility.com/D8764