[dagit] Fix link to partition from schedule page
Summary:

The link from a schedule details section to a relevant partition set does not currently include the `mode`, which means it can end up on the wrong job permalink. If the job doesn't have any partitions, the user just ends up on the overview tab for that pipeline, which isn't useful anyway.

Make sure the `mode` is present, which will result in landing on the correct tab (Partitions) and the correct partition.

Test Plan:

View a schedule in/out of crag flag, verify that the link correctly includes `pipeline:mode` in both cases.

Reviewers:

bengotow, dgibson, sandyryza

Subscribers: