Add dbt pipeline to example repo, synchronize with internal
Summary:

After this commit lands, this will be the source of truth for the current state of the hacker_news demo. A separate commit will strip down the internal repo to an import of this, along with some band-aid patches where necessary (e.g. links to specific demo sites)

Test Plan: bk

Reviewers: cdecarolis, yuhan, sandyryza

Subscribers: