[dagit] Handle WS issue in CSP for Safari
Summary:

The `connect-src: 'self'` configuration in the CSP is failing in Safari, which is rejecting WebSocket connections. Add `ws:` and `wss:`, which is really unfortunate. I'd like to fix Safari temporarily with this, but would really prefer a better long-term solution than this.

Test Plan:

Test dev and prod builds in Chrome, Firefox, Safari. Verify that Dagit loads and performs WS connections correctly.

Reviewers:

bengotow, dgibson, max

Subscribers: