[crag] Add fields not brought about in default config, fix bug with execute_in_process and default config, catch bad config earlier on .
Summary: Currently, we aren't carrying over fields from the original config if they aren't specified in the default config for a job. This leads to either weird errors, or wrong config in some cases. This attempts to fix the issue.

That fix also required me to finagle around the executor config that is now always carried over.

Test Plan: unit tests

Reviewers: sandyryza, alangenfeld

Subscribers: