Use grpc rather than ipc to decide if a grpc server is running
Summary:
A user is reporting that writing to a tempdir on their machine takes 6 minutes. This is a nice excuse to clean up the way we do gRPC server startup - we have a perfectly good way to check if a server is running, so let's use it.

The one thing that we lose here is reporting a load error on startup in a structured way, but the --lazy-load-user-code

Test Plan: BK, run dagit locally, gRPC servers still load, simulate load failure, get a clear error in dagit

Reviewers: alangenfeld, max

Differential Revision: https://dagster.phacility.com/D9046