[memoization 7/n] retrieve memoized results from execute_in_process
Summary: Something users have run into when using memoization is difficulty actually getting visibility into their run results. Once their pipeline is stable, and they run it, they might end up in a situation where it's onerous to retrieve output results (which it's probably reasonable to want to do during development). This makes it so that we can look back through memoized outputs if requested.

We will probably also want to support this from the submit API, but I think based on the fact that we'll always be relying on the IO manager directly for outputs, it will be even easier.

Test Plan: added a test to ensure that this is possible.

Reviewers: sandyryza, alangenfeld

Subscribers: