Add an error boundary within get_pipeline_run_observable to prevent dagit hanging on message parsing failures
Summary: For reasons that we're honestly not entirely sure of, raising an exception within this event parsing code doesn't give the postgres event listener a chance to clean up, leaving it in a hanging state. This is an attempt to add an error boundary to avoid uncaught exceptions within the Observable callback.

Test Plan: Simulate an error within the event parsing code, see that the failure no longer bricks dagit

Reviewers: alangenfeld, prha

Subscribers: dish

Differential Revision: https://dagster.phacility.com/D9182